projectile collision with OOF

+++ jeu/src/game.cpp
@@ -1,10 +1,10 @@
 /**
  * TODO:
  *  thread pour jouer en même temps
- *  projectile collision
- *  ERROR:
+ *  enlever static_cast !!
+ * ERROR:
  *  shoot => out of range
- *  hp doesnt change
+
  */
 
 #include <unistd.h>
@@ -58,7 +58,7 @@ int init() {
     main_wnd = newwin(screen_area.height(), screen_area.width(), 0, 0);
 
     // define area for movement
-    game_area = { {0, 0}, {screen_area.width() - 2, screen_area.height() - infopanel_height - 4}};
+    game_area = { {0, 0}, {static_cast<uint_fast16_t>(screen_area.width() - 2), static_cast<uint_fast16_t>(screen_area.height() - infopanel_height - 4)}};
 
     init_pair(1, COLOR_WHITE, COLOR_BLACK);
     init_pair(2, COLOR_GREEN, COLOR_BLACK);
